Rewrite pkg_source
Old script was sort of broken and jank but it has been improved
Fixed possible issue in src/lexer.c with the structure member `lexer_scopes::data`

+++ src/lexer.c
@@ -22,8 +22,7 @@ enum lexer_type {
 };
 
 struct lexer_scopes {
-	uint16_t data;
-
+	uint32_t data;
 	union {
 		struct {
 			uint16_t param;
